checking lot owner in api

+++ app/controllers/ApiDeleteLotController.php
@@ -2,6 +2,8 @@
 namespace App\Controllers;
 
 use App\Models\ApiLotManipulate;
+use App\Models\ApiUserGet;
+use App\Models\ApiLotValidate;
 
 class ApiDeleteLotController
 {  
@@ -13,11 +15,20 @@ class ApiDeleteLotController
 
     public static function apiDeleteLot(int $lot_id)
     {
-        $lot = new ApiLotManipulate();
+        $userId = (new ApiUserGet)->getIdViaToken($_COOKIE['api_token']);
 
-        header('HTTP/1.0 201');
-        header('Content-Type: application/json; charset=UTF-8');
-    
-        echo $lot->deleteLot($lot_id);
+        if ((new ApiLotValidate)->checkOwnerId($userId, $lot_id)) {
+            $lot = new ApiLotManipulate();
+
+            header('HTTP/1.0 201');
+            header('Content-Type: application/json; charset=UTF-8');
+        
+            echo $lot->deleteLot($lot_id);
+        } else {
+            header('HTTP/1.0 201');
+            header('Content-Type: application/json; charset=UTF-8');
+
+            echo json_encode(['Error!' => 'This isn\'t youre lot'], JSON_UNESCAPED_UNICODE);
+        }
     }
 }
